@@ -1,26 +1,21 @@
 import 'package:amparo_app/screens/pages/historypage_screen.dart';
 import 'package:flutter/material.dart';
 
-// Stateful é para modificar a tela, por exemplo: o teclado sobe
-class LoginScreen extends StatelessWidget {
+class Login extends StatelessWidget {
+  final bool _showPassword = false;
+  final String asylumName;
 
-  final String nomeAsilo;
-  String nomeUsuario;
-  LoginScreen({Key key, @required this.nomeAsilo, this.nomeUsuario}) : super(key: key);
+  Login({required this.asylumName});
 
-  bool _showPassword = false;
-
-  void _enviaDados(BuildContext context){
-
+  void _sendData(BuildContext context){
     Navigator.push(
         context,
-        MaterialPageRoute(builder: (context) => HistoryPage(nomeAsilo: nomeAsilo, nomeUsuario: nomeUsuario)));
+        MaterialPageRoute(builder: (context) =>
+            HistoryPage(nomeAsilo: nomeAsilo, nomeUsuario: nomeUsuario)));
   }
 
   @override
   Widget build(BuildContext context) {
-
-
     return Scaffold(
       appBar: AppBar(
         elevation: 0,
@@ -52,18 +47,14 @@ class LoginScreen extends StatelessWidget {
       ),
       backgroundColor: Color(0xFF1D6AFF),
       body: Column(
-
         children: <Widget>[
-
           Expanded(
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
               mainAxisSize: MainAxisSize.min,
-
               children: <Widget>[
-
                 Text(
-                    'Bem-vindo ao ' + nomeAsilo,
+                    'Bem-vindo ao ' + asylumName,
                     textAlign: TextAlign.center,
                     style: TextStyle(
                       fontSize: 34,
@@ -85,9 +76,7 @@ class LoginScreen extends StatelessWidget {
               ],
             ),
           ),
-
           Container(
-
             child: Card(
               margin:  const EdgeInsets.symmetric(horizontal: 0),
               shape: RoundedRectangleBorder(
@@ -95,13 +84,10 @@ class LoginScreen extends StatelessWidget {
                       topLeft: Radius.circular(50), topRight: Radius.circular(50)
                   )
               ),
-
               child: Padding(
                 padding: const EdgeInsets.all(16),
-
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.stretch,
-
                   children: [
                     Text(
                       'Acesse sua conta',
@@ -113,11 +99,8 @@ class LoginScreen extends StatelessWidget {
                         color: Colors.black,
                       ),
                     ),
-
-
                     Padding(
                       padding: const EdgeInsets.only(left: 3, bottom: 4, top: 6),
-
                       child: Text(
                         'Caso não possua credenciais de acesso, contate o administrador',
                         textAlign: TextAlign.center,
@@ -128,10 +111,8 @@ class LoginScreen extends StatelessWidget {
                         ),
                       ),
                     ),
-
                     Padding(
                       padding: const EdgeInsets.only(left: 10,  top: 0, right: 10),
-
                       child: TextFormField(
                         decoration: InputDecoration(
                           icon: Icon(Icons.person),
@@ -151,10 +132,8 @@ class LoginScreen extends StatelessWidget {
                         },
                       ),
                     ),
-
                     Padding(
                       padding: const EdgeInsets.only(left: 10,  top: 10, right: 10),
-
                       child: TextFormField(
                         decoration: InputDecoration(
                           icon: Icon(Icons.lock),
@@ -175,13 +154,10 @@ class LoginScreen extends StatelessWidget {
                         ),
                         keyboardType: TextInputType.text,
                         obscureText: true,
-
                       ),
                     ),
-
                     Padding(
                       padding: const EdgeInsets.only(left: 10, top: 10, right: 10),
-
                       child: GestureDetector(
                         child: Text(
                           'Esqueceu a senha?',
@@ -194,10 +170,8 @@ class LoginScreen extends StatelessWidget {
                         ),
                       ),
                     ),
-
                     Padding(
                       padding: const EdgeInsets.only(left: 10, bottom: 4, top: 10, right: 10),
-
                       child: RaisedButton(
                         color: Color(0xFF1D6AFF),
                         child: Text(
@@ -212,23 +186,14 @@ class LoginScreen extends StatelessWidget {
                         shape: new RoundedRectangleBorder(
                             borderRadius: BorderRadius.circular(30),
                             side: BorderSide(color: Colors.blueAccent)
-
                         ),
-                        onPressed: (){
-                          _enviaDados(context);
-                        },
+                        onPressed: () => Navigator.push(context, MaterialPageRoute(builder: (context) => ResidentList())),
                       ),
-                    ),
-
-                  ],
+                    ),],
                 ),
               ),
-
             ),
-
           )
-
-
         ],
       ),
     );
